% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rail_function.R
\name{rail_function}
\alias{rail_function}
\title{A Rail Function}
\usage{
rail_function(name, love = TRUE)
}
\arguments{
\item{name}{- enter your name e.g. "Auriel"}

\item{love}{Do you love rails? Defaults to TRUE.}
}
\description{
This function allows you to express your love of rails
}
\examples{
rail_function("Auriel",love=TRUE)
}
\keyword{rails}

